Safari-611.1.10/LayoutTests/js/var-shadows-arg-crash-expected.txt

Tests to ensure that activations are built correctly in the face of duplicate parameter names and do not cause crashes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test()() is "success"
PASS successfullyParsed is true

TEST COMPLETE
